Splits list-templates command from new --list.
This will let us share the list-templates command with other projects and the CLI plugin more easily.
